When API request is unsuccessful
Added a for loop and if statement to check the API response. It loops through the data dictionary, and will alert the user if there is an error from the request. I have included a prompt to check that API creds have been included as this will be the most likely issue.