Improved flow for checking the data received and tidied to search_again function
Added a recipes_found variable.
This changes the final message when the user exits the program depending on whether their API request(s) got recipe results.

Added a response_code variable. This meant the for loop was no longer needed to match the key with certain values in the data dictionary.

Improved the search_again function to reduce duplicate lines of code.