Fixed error for 401 response
Added a response_code variable to the global scope. This means that only when a 401 error is thrown will the user be prompted to re-enter API creds. I also fixed an issue where an error was thrown after confirming the creds.

Also removed an errant print(data) from the case where recipes are found.