Try and except statement
Added try and except statement to improve error handling when a user types in a dietary requirement that does not match one of the keys in the 'options' dictionary. When there is a KeyError, a print statement directs the user to matching their requirement to one of the options available, and the loop continues.